Gjer sletteknapp for huskeliste gjenbrukbar
slik at vi kan bruke den både i visning og redigering utan å repetere logikken for om brukaren har lov til å slette huskelappen.
