+++ src/component/huskelapp/visning/huskelapp-visning-modal.tsx
@@ -1,16 +1,15 @@
 import './huskelapp-visning.less';
 import { useModalStore } from '../../../store/modal-store';
 import { BodyShort, Button, Detail, Heading, Modal } from '@navikt/ds-react';
-import { TrashIcon } from '@navikt/aksel-icons';
 import HuskelappIkon from '../ikon/huskelapp.svg?react';
 import { useDataStore } from '../../../store/data-store';
 import { toSimpleDateStr } from '../../../util/date-utils';
 import { trackAmplitude } from '../../../amplitude/amplitude';
-import { kanFjerneHuskelapp } from '../../../util/selectors';
+import { SlettHuskelapp } from '../redigering/slett-huskelapp';
 
 function HuskelappVisningModal() {
-    const { hideModal, showHuskelappRedigereModal, showFjernHuskelappModal } = useModalStore();
-    const { huskelapp, innloggetVeileder, oppfolging } = useDataStore();
+    const { hideModal, showHuskelappRedigereModal } = useModalStore();
+    const { huskelapp } = useDataStore();
 
     const endreHuskelappKlikk = () => {
         trackAmplitude({
@@ -21,21 +20,14 @@ function HuskelappVisningModal() {
     };
 
     /* TODO Burde vi ha noko ordentleg feilhandtering her i staden for berre å ikkje vise noko
-     *   dersom vi manglar huskelapp eller innlogga veileder?
+     *   dersom vi manglar huskelapp?
      *  Dette skal i teorien ikkje kunne skje fordi vi sjekkar kva modal som skal visast basert på
-     *   om vi har huskelapp eller ikkje og om innlogga brukar jobbar på same kontor som oppfølgande veileder. */
+     *   om vi har huskelapp eller ikkje  */
     if (!huskelapp) {
         // eslint-disable-next-line no-console
         console.warn('Oisann, her prøver noen å åpne huskelapp uten at det finnes noen huskelapp å åpne.');
         return null;
     }
-    if (!innloggetVeileder) {
-        // eslint-disable-next-line no-console
-        console.warn('Oisann, her prøver noen å åpne huskelapp uten at vi får tak i innlogget veileder.');
-        return null;
-    }
-
-    const veilederKanSletteHuskelapp = kanFjerneHuskelapp(huskelapp, oppfolging, innloggetVeileder.ident);
 
     return (
         <Modal
@@ -67,17 +59,7 @@ function HuskelappVisningModal() {
                 <Button onClick={endreHuskelappKlikk} size="small" variant="primary" form="huskelapp-form">
                     Endre
                 </Button>
-                {veilederKanSletteHuskelapp && (
-                    <Button
-                        onClick={showFjernHuskelappModal}
-                        icon={<TrashIcon aria-hidden />}
-                        size="small"
-                        variant="secondary"
-                        type="button"
-                    >
-                        Slett
-                    </Button>
-                )}
+                <SlettHuskelapp />
             </Modal.Footer>
         </Modal>
     );
